updated the codebase to use OpenMP parallelism
