+++ CMakeLists.txt
@@ -5,6 +5,20 @@ project("Planck")
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
+# Set optimization flags
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3")
+
+# Option to enable/disable OpenMP support
+option(ENABLE_OPENMP "Enable OpenMP support" OFF)
+
+# Check if OpenMP is requested and available
+if(ENABLE_OPENMP)
+    find_package(OpenMP REQUIRED)
+    if(OpenMP_CXX_FOUND)
+        set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${OpenMP_CXX_FLAGS}")
+    endif()
+endif()
+
 # Check if Eigen is available, otherwise download and build it
 find_package(Eigen3 QUIET)
 if(NOT Eigen3_FOUND)
@@ -20,7 +34,7 @@ if(NOT Eigen3_FOUND)
         INSTALL_COMMAND ""
     )
     ExternalProject_Get_Property(Eigen SOURCE_DIR)
-    set(EIGEN3_INCLUDE_DIRS ${SOURCE_DIR})
+    set(EIGEN3_INCLUDE_DIR ${SOURCE_DIR})
 endif()
 
 # Check if Boost is available, otherwise download and build it
@@ -39,21 +53,39 @@ if(NOT Boost_FOUND)
         INSTALL_COMMAND ""
     )
     ExternalProject_Get_Property(Boost SOURCE_DIR)
-    set(BOOST_INCLUDE_DIRS ${SOURCE_DIR})
+    set(BOOST_INCLUDE_DIR ${SOURCE_DIR})
 endif()
 
-# Add source files
-file(GLOB EXECUTABLE_SOURCES "src/*.cpp")
+# Add the executable target
+add_executable(planck src/planck.cpp)
 
-# Create an executable for each source file
-foreach(EXECUTABLE_SOURCE ${EXECUTABLE_SOURCES})
-    get_filename_component(EXECUTABLE_NAME ${EXECUTABLE_SOURCE} NAME_WE)
-    add_executable(${EXECUTABLE_NAME} ${EXECUTABLE_SOURCE})
-    # Link libraries
-    target_link_libraries(${EXECUTABLE_NAME} Eigen3::Eigen ${Boost_LIBRARIES})
-    # Set output directory for binary
-    set_target_properties(${EXECUTABLE_NAME} PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin)
-endforeach()
+# Include directories
+target_include_directories(planck PUBLIC 
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/auxiliary
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/molecule 
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/scf
+)
+
+# Add source files from auxiliary directory using file globbing
+file(GLOB AuxiliarySources src/auxiliary/*.cpp)
+
+# Add source files from molecule directory using file globbing
+file(GLOB MoleculeSources src/molecule/*.cpp)
+
+# Add source files from scf directory using file globbing
+file(GLOB ScfSources src/scf/*.cpp)
+
+# Add source files to the target
+target_sources(planck PRIVATE
+    ${MoleculeSources}
+    ${ScfSources}
+    ${AuxiliarySources}
+)
+
+# Set the output directory for the planck executable
+set_target_properties(planck PROPERTIES
+    RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
+)
 
 # Copy the basis set directory to the bin folder
 file(COPY ${CMAKE_SOURCE_DIR}/basis DESTINATION ${CMAKE_BINARY_DIR}/bin)
