+++ src/auxiliary/tables.cpp
@@ -1,11 +1,4 @@
-#pragma once
-
-#include <cmath>
-#include <fstream>
-#include <iostream>
-#include <map>
-#include <string>
-#include <vector>
+#include "tables.h"
 
 /*-----------------------------------------------------------------------------
  * Planck
@@ -23,13 +16,6 @@
  * You should have received a copy of the GNU General Public License along with
  * this program.  If not, see <http://www.gnu.org/licenses/>.
  ----------------------------------------------------------------------------*/
-
-/*
- * This is a lookup table for atomic masses, atomic numbers and atomic radii for
- * elements from Hydrogen to Einsteinium.
- *
- * Data obtained from PubChem
- */
 
 std::map<std::string, double> atomicMass = {
     {"H", 1.008}, {"He", 4.003}, {"Li", 7.000}, {"Be", 9.012}, {"B", 10.810}, {"C", 12.011}, {"N", 14.007}, {"O", 15.999}, {"F", 18.998}, {"Ne", 20.180}, {"Na", 22.990}, {"Mg", 24.305}, {"Al", 26.982}, {"Si", 28.085}, {"P", 30.974}, {"S", 32.070}, {"Cl", 35.450}, {"Ar", 39.900}, {"K", 39.098}, {"Ca", 40.080}, {"Sc", 44.956}, {"Ti", 47.867}, {"V", 50.942}, {"Cr", 51.996}, {"Mn", 54.938}, {"Fe", 55.840}, {"Co", 58.933}, {"Ni", 58.693}, {"Cu", 63.550}, {"Zn", 65.400}, {"Ga", 69.723}, {"Ge", 72.630}, {"As", 74.922}, {"Se", 78.970}, {"Br", 79.900}, {"Kr", 83.800}, {"Rb", 85.468}, {"Sr", 87.620}, {"Y", 88.906}, {"Zr", 91.220}, {"Nb", 92.906}, {"Mo", 95.950}, {"Tc", 96.906}, {"Ru", 101.100}, {"Rh", 102.906}, {"Pd", 106.420}, {"Ag", 107.868}, {"Cd", 112.410}, {"In", 114.818}, {"Sn", 118.710}, {"Sb", 121.760}, {"Te", 127.600}, {"I", 126.905}, {"Xe", 131.290}, {"Cs", 132.905}, {"Ba", 137.330}, {"La", 138.906}, {"Ce", 140.116}, {"Pr", 140.908}, {"Nd", 144.240}, {"Pm", 144.913}, {"Sm", 150.400}, {"Eu", 151.964}, {"Gd", 157.200}, {"Tb", 158.925}, {"Dy", 162.500}, {"Ho", 164.930}, {"Er", 167.260}, {"Tm", 168.934}, {"Yb", 173.050}, {"Lu", 174.967}, {"Hf", 178.490}, {"Ta", 180.948}, {"W", 183.840}, {"Re", 186.207}, {"Os", 190.200}, {"Ir", 192.220}, {"Pt", 195.080}, {"Au", 196.967}, {"Hg", 200.590}, {"Tl", 204.383}, {"Pb", 207.000}, {"Bi", 208.980}, {"Po", 208.982}, {"At", 209.987}, {"Rn", 222.018}, {"Fr", 223.020}, {"Ra", 226.025}, {"Ac", 227.028}, {"Th", 232.038}, {"Pa", 231.036}, {"U", 238.029}, {"Np", 237.048}, {"Pu", 244.064}, {"Am", 243.061}, {"Cm", 247.070}, {"Bk", 247.070}, {"Cf", 251.080}, {"Es", 252.083}};
