--- src/planck
@@ -1,1 +1,0 @@
-./planck.sh
